Fixed warnings

+++ libabclient/libabclient/PropStream.h
@@ -50,7 +50,7 @@ public:
         if (GetSize() < len)
             return false;
 
-        char* str = new char[len + 1];
+        char* str = new char[static_cast<size_t>(len) + 1];
         memcpy_s(str, len, p_, len);
         str[len] = 0;
         r.assign(str, len);
